+++ lib/Widgets/Home/MultiCallButton.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:meetby_accident/Widgets/Home/GiantButton.dart';
+import 'package:meetby_accident/Widgets/Reuse/UserWidgets.dart';
+import 'package:meetby_accident/other/datahandler/Userhelper.dart';
 
 class MultiCallButton extends StatelessWidget {
   @override
@@ -10,21 +12,52 @@ class MultiCallButton extends StatelessWidget {
       child: GiantButton(
         child: Padding(
           padding: EdgeInsets.all(18.0),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          child: Column(
             children: <Widget>[
-              Text(
-                "Zufällig mehrere Kontakte anrufen",
-                style: TextStyle(
-                  fontSize: 16,
-                  color: Theme.of(context).textTheme.headline1.color,
-                ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: <Widget>[
+                  Text(
+                    "Zufällig mehrere Kontakte anrufen",
+                    style: TextStyle(
+                      fontSize: 16,
+                      color: Theme.of(context).textTheme.headline1.color,
+                    ),
+                  ),
+                  Icon(
+                    Icons.call,
+                    color: Colors.green,
+                  ),
+                ],
               ),
-              Icon(Icons.call, color: Colors.green,),
+              SizedBox(height: 20),
+              Roller(),
             ],
           ),
         ),
       ),
+    );
+  }
+}
+
+class Roller extends StatefulWidget {
+  @override
+  _RollerState createState() => _RollerState();
+}
+
+class _RollerState extends State<Roller> {
+  Widget users = Row(
+    children: <Widget>[
+      ProfileIcon(
+        UserHelper().random,
+      ),
+    ],
+  );
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: users,
     );
   }
 }
